pyproject.toml
[project]
name = "appstore_data_agent"
version = "0.1.0"
description = "appstore_data_agent using crewAI"
authors = [{ name = "Viral Rathod", email = "rathod.viral@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "argparse>=1.4.0",
    "bs4>=0.0.2",
    "crewai[tools]>=0.177.0",
    "googlesearch-python>=1.2.0",
    "ollama>=0.5.3",
    "pandas>=2.2.0",
    "pytest>=8.4.1",
    "requests>=2.32.5",
]

[project.scripts]
appstore_data_agent = "appstore_data_agent.main:run"
run_crew = "appstore_data_agent.main:run"
train = "appstore_data_agent.main:train"
replay = "appstore_data_agent.main:replay"
test = "appstore_data_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
